+++ HISTORY.md
@@ -11,6 +11,7 @@
 * Moved include/utilities/*.h to include/rocksdb/utilities/*.h
 * Statistics APIs now take uint32_t as type instead of Tickers. Also make two access functions getTickerCount and histogramData const
 * Add DB property rocksdb.estimate-num-keys, estimated number of live keys in DB.
+* Add DB::GetIntProperty(), which returns DB properties that are integer as uint64_t.
 
 ## 3.3.0 (7/10/2014)
 ### New Features